rename main package
- remove mention to fr.edgewhere
- make it a possibly public repository from com.cyrildever
