+++ src/main/scala/com/cyrildever/feistel/client/Operation.scala
@@ -1,0 +1,13 @@
+package com.cyrildever.feistel.client
+
+/**
+ * Operation class
+ *
+ * @author  Cyril Dever
+ * @since   1.0
+ * @version 2.0
+ */
+object Operation {
+  val ENCRYPT = "encrypt"
+  val DECRYPT = "decrypt"
+}
