--- src/main/scala/fr/edgewhere/feistel/client/Operation.scala
@@ -1,13 +1,0 @@
-package fr.edgewhere.feistel.client
-
-/**
- * Operation class
- *
- * @author  Cyril Dever
- * @since   1.0
- * @version 1.0
- */
-object Operation {
-  val ENCRYPT = "encrypt"
-  val DECRYPT = "decrypt"
-}
