+++ src/test/scala/com/cyrildever/BasicUnitSpecs.scala
@@ -1,0 +1,13 @@
+package com.cyrildever
+
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+/**
+ * BasicUnitSpecs class
+ *
+ * @author      Cyril Dever
+ * @version     3.0
+ * @since       1.0
+ */
+abstract class BasicUnitSpecs extends AnyFlatSpec with Matchers
