--- src/test/scala/fr/edgewhere/BasicUnitSpecs.scala
@@ -1,13 +1,0 @@
-package fr.edgewhere
-
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers
-
-/**
- * BasicUnitSpecs class
- *
- * @author      Cyril Dever
- * @version     2.0
- * @since       1.0
- */
-abstract class BasicUnitSpecs extends AnyFlatSpec with Matchers
